Refactor - Adjust rating to float value

+++ db/migrate/20211015005626_create_food_groups.rb
@@ -2,7 +2,7 @@ class CreateFoodGroups < ActiveRecord::Migration[5.2]
   def change
     create_table :food_groups do |t|
       t.string :name
-      t.integer :rating_out_of_ten
+      t.float :rating_out_of_ten
       t.boolean :perishable
 
       t.timestamps
